Page tables: unmap_page(): fix call to range handler
The range passed to the range handler should be the same range that
has just been unmapped.
This change fixes an incorrect deallocation of physical memory that
happens when a memory range being unmapped contains a 2MB-sized
page that covers a portion of the tail of the unampped range (and
that is thus split into 4KB pages); see the "bitmap_dealloc error"
messages in https://github.com/nanovms/ops/issues/1654.