Filesystem: retire fs_status enum in favor of standard error codes
Having filesystem-specific error codes requires conversion between
filesystem codes and POSIX standard codes every time a filesystem-
related error is reported to userspace. For the 9P filesystem, in
some cases a double conversion (from a standard code to a
filesystem code and then back to a standard code) may need to be
done.
This change removes the fs_status enum that declares the filesystem
error codes, and uses standard POSIX codes to report filesystem
errors; this allows eliminating the above conversions. Standard
error codes are defined in the new kernel/errno.h header file; in
order to prevent this file from being included from userspace code,
the kernel directory is being removed from the include paths in
the relevant Makefiles; this required some adjustments to other
header files so that they can be included by both kernel and non-
kernel code.
For the 9P filesystem, the p9_create() callback function is being
modified so that EOPNOTSUPP is returned instead or EINVAL if the
user program tries to open a file with the O_TMPFILE flag (#2051).