pagecache_read_sg_finish(): fix locking order inversion
Throughout the pagecache code, when both a cache node and the cache
state need to be locked at the same time, the node must be locked
first. The existing implementation of pagecache_read_sg_finish() is
locking the cache state first, thereby creating a locking order
inversion that can cause the kernel to deadlock in multi-vCPU
instances, for example if multiple reads from the page cache are
executed concurrently (the CI test failure at
https://app.circleci.com/pipelines/github/nanovms/nanos/4810/workflows/d6be64e2-18b7-4d4a-bb4c-8426ac35239c/jobs/16820
is caused by this).
Bug introduced in commit 8225e715.