TFS extent extension: fix handling of exhausted disk space
If there is not enough storage space to extend a TFS file extent to
the requested size, the extent size should be adjusted to match the
available space. The existing code is updating the extent size as
if the storage space limit had not been reached; this can cause the
TFS driver to issue disk I/O requests with invalid block ranges,
which results in EIO syscall errors.