Runloop: fix check for unset CPU timer when running a user thread
Before running a user thread, if the local timer of the current CPU
is not armed (because either the timer interrupt already fired
since doing the last timer update, or the timer has never been
armed), the `timeout` variable in the current code in
runloop_internal() is negative, thus the
`timeout > (s64)max_timeout` condition is always false; this
results in the CPU timer not being armed and the user thread
potentially running indefinitely.
This issue causes the umcg runtime test to hang if run on an
instance with more than 4 vCPUs.
This change fixes the above issue by using the `last_timer_update`
field of the cpuinfo struct to check the timestamp at which the
next timer interrupt will fire (with a 0 value in this field
indicating that the timer is not armed). In addition, unnecessary
updates of the CPU timer are avoided by removing the call to
`set_platform_timer` from `update_timer` (because the timer may
have to be re-configured before kernel_sleep() is finally called),
and checking whether the current timer setting is different from
the next timer timestamp before updating the timer setting.
The `empty` field of struct timerqueue is being removed since it is
no longer used.