Makefiles: consolidate setting of clang `-target` flag
The clang compiler needs to be told (for any architecture if on
MacOS, only when cross-compiling when on Linux) the compilation
target. This change adds setting of the relevant `-target` flag in
a common file (rules.mk) and removes this setting from duplicated
places; in addition, this flag is being added where appropriate in
the ARM and RiscV architectures, in preparation for supporting
building via clang in those architectures.