Makefile: add `kernel` make target
This new make target builds the kernel, the bootloaders (if
applicable for a given platform), and the klibs. It will allow the
core pieces of a unikernel image to be built without the compiler
toolchain that is needed to build user-mode programs for the
target architecture.
The `release` target in the top Makefile has been updated to not
include the mkfs and dump tools in the release tarball (they are no
longer needed since Ops implements internally equivalent
functionalities), and the platform-specific `release` targets have
been removed since they are obsoleted by the target in the top
Makefile.
The `make tools` command has been added to the build-pc CircleCI
job because the dump tool is no longer built by the default make
target, and this tool is needed to run the extended tests (more
specifically, the core dump test).