unix_heaps struct: remove embedded kernel_heaps struct
The initialization of the empedded struct in init_unix() is making
the clang compiler produce a call to memcpy(), which is not an
available function for the kernel code.
This change removes the embedded kernel_heaps struct from the
unix_heaps struct, since a pointer to the kernel_heaps can be
easily obtained via get_kernel_heaps().