unix/mmap.c: remove copying of struct vmap
The vmap struct passed by value to allocate_vmap_locked() and
returned as value by alter_vmap_key() is making the clang compiler
produce calls to memcpy(), which is not an available function for
the kernel code.
This change modifies the above functions so that copying of vmap
structs is avoided.