Klibs: move build artifacts to output/platform/<PLATFORM> folder
This change allows building the klibs for different target
platforms from the same source tree without having to purge the
output/klib directory between builds.
Some modifications have been made to a number of Makefiles in order
to be able to do `make clean` (which deletes the build artifacts
for a single target platform) when the filesystem contains build
artifacts for multiple platforms.