Process stack: add guard gap from adjacent mapping
When a user process overflows its stack, the kernel is expected to
generate a SIGSEGV; however, if there is an adjacent mapping
contiguous to the stack mapping, the page fault that is needed in
order to trigger the SIGSEGV may not occur, and as a result the
stack overflow may not be detected.
This change adds a guard gap (with a size of 256 pages, i.e. 1MB)
between the process stack and any adjacent mapping; this greatly
decreases the likelihood of a stack overflow not being detected,
and prevents the exploitation of many application vulnerabilities
thar involve overflowing the stack.