Unix filesystem: add file_mode_from_type() utility function
This new function is now used by the stat, lstat, fstat, and
newfstatat syscalls; in the next commit, it will also be used by
the statx syscall.
In addition, for symbolic links, the st_size field of the stat
structure is now correctly set to the length of the link target
path.