pagecache_read_sg(): allow use with non-contextual closure
The existing code is assuming that the completion passed to
pagecache_read_sg() is a contextual closure. This prevents reading
from the page cache with a completion that is not a contextual
closure, such as when the reading context suspends itself and the
completion resumes the suspended context (this happens in the
Nvidia GPU klib when reading the GPU firmware file).
This change allows reading from the page cache with non-contextual
completion closures.