Page faults: add support for custom fault handlers
This allows custom mmap() implementations for specific file
descriptors to handle page faults on mmap()ed memory. This feature
is used by the Nvidia GPU klib.
The new_pending_fault_locked() function is no longer static so that
it can be called from an arbitrary fault handler when a fault
cannot be resolved synchronously.