Sockets: add support for asynchronous read/write
Socket read/write operations can be trigggered asynchronously, e.g.
via io_uring; thus, the context to be associated to the blockq for
a read/write operation may not be the current context; instead,
this context must be retrieved from the closure that is created
when triggering the operation.
In addition, in the blockq_check_timeout() function the
`blocked_on` field of the unix_context struct must be set
regardless of whether the operation is synchronous.