Aarch64 powerdown: add support for GPIO power key
If ACPI tables are not present in a given Qemu aarch64 instance
(e.g. when running an instance with the "virt" machine type and
without a UEFI firmware), upon reception of the "system_powerdown"
command (which is used to gracefully shut down the instance), Qemu
toggles the "poweroff" GPIO instead of generating an ACPI event.
Add parsing of GPIO key information from the device tree, and if
such information can be found, enable interrupt generation on the
PL061 GPIO controller when the poweroff GPIO is toggled, so that
the kernel can be shut down by the interrupt handler.
This allows a graceful shutdown to be performed when an on-prem
instance launched with `ops run --arch=arm64` is terminated with
CTRL-C.