ARM physical heap: exclude all identity-mapped memory
The memory range that is identity-mapped during boot is used, among
other things, as temporary stack for both the boot processor and
the secondary processors; theefore, it should not be included in
the physical memory ranges, otherwise a secondary processor can
corrupt memory that has been allocated for other kernel uses.
This fixes unhandled exceptions that cause boot failures when
running the kernel on Azure Standard_D2ps_v5 instances.