Arm is_illegal_instruction(): add check for trapped instructions
When the user process executes MSR, MRS or System instructions, the
processor raises an exception, with the exception code (EC) in the
Exception Syndrome Register (ESR_EL1) set to 0x18 (trapped MSR,
MRS or System instruction execution in AArch64 state), which should
be treated as an illegal instruction and cause the kernel to send
a SIGILL signal to the process.
Amend the Arm implementation of is_illegal_instruction() so that it
recognizes this type of exceptions.