Arm unix_fault_handler(): add emulation of MRS instruction
Starting with Go version 1.23, the Golang runtime for Arm64 during
initialization tries to retrieve the values of some system
registers (via the MRS instruction) in order to evaluate the
features implemented by the processor. This instruction causes a
CPU exception, which the kernel handles by sending SIGILL to the
user process; as a result, Go programs built with Go version 1.23
do to not work on Arm.
Add emulation of the MRS instruction to retrieve the values of the
system registers used by the Go runtime, so that Go 1.23 works
properly on Arm.