Makefile: add support for generic kernel configuration options
In order to allow kernel configuration options to be specified in a
make command, add handling of a KCONFIG Makefile variable, which
creates a kconfig.h header file that is included by all kernel
source files. The KCONFIG variable can contain an arbitrary number
of comma-separated options, which are converted into #defines in
the header file; during the conversion, `-` characters are replaced
with `_`, and lowercase characters are converted to uppercase. For
example `KCONFIG=opt1,no-opt2` will result in the following
kconfig.h file:
```
#define OPT1
#define NO_OPT2
```

Integrate handling of the existing debug options into the new
kconfig.h file, which replaces the old debug.h file (and is placed
in the platform-specific output folder instead of the common output
folder, so that the build process can correctly determine when the
kernel configuration changes for a given platform and thus when the
kernel needs to be rebuilt).