KASLR: add compile-time option to disable randomization
A kernel built with KASLR disabled is more debug-friendly.
Edit the gdb initialization files so that .gdbinit contains only
commands common to all architectures, while the .gdbinit-* files
contain architecture-specific commands (where the kernel symbol
offset is set to the value used when KASLR is disabled) and are
meant to be used via the `-ix` gdb command line option.