runloop: fix servicing of kernel timers
When the update_timer() function is called during the runloop, if
it returns a non-zero timstamp, the returned value must be used to
trigger an interrupt on the local timer, otherwise the relevant
timer event is missed. If the aync queues are serviced after the
update_timer() call, when an async operation does not return the
timestamp returned by update_timer() is lost and the local timer
is not armed, thereby missing the relevant timer event.
Fix this issue by moving all servicing of aync queues at the
beginning of the runlopp, before the update_timer() call, using a
loop that runs until no operations are dequeued from any aync
queue.