Memory heap: use MEM_NOFAIL flag for boot-time allocations
For memory allocations done at boot time that are required for
starting the kernel (including klibs if any) and the user program,
replace calls to the allocate() macro with calls to the mem_alloc()
function (with the MEM_NOFAIL flag set): this allows removing
relevant assert() statements and therefore decreases code size. Use
the additional MEM_NOWAIT flag for allocations done before (or
during) kernel_runtime_init(), because at that time the
functionality to suspend a context or wait for other CPUS is not in
place yet.
Do the same change also for allocations that are not supposed to
fail in user programs (unit and runtime tests).