file_readahead(): use asynchronous task to fetch pages
This function is called from a syscall context after instantiating
a contextual closure that resumes the context upon file I/O
completion. If this function suspends the current context while
waiting for pages to be fetched from the cache, the contextual
closure can be invoked during context suspension, which corrupts
the stack of the suspended context and causes a kernel crash when
the context is resumed. Fix this issue by using an asynchronous
task (whose context can be safely suspended) to fetch the pages.