VirtIO balloon: enhance memory cleaner implementation
The memory cleaner for the virtio balloon device deflates the
balloon when receiving a request to release memory.
To follow the virtio specifications, instantiate the cleaner only
when the "deflate_on_oom" feature is negotiated, and when the
cleaner is instantiated, deflate the balloon only when the system
is running out of memory.
In addition, add support for deflating the balloon synchronously
(i.e. suspending the current context to wait for the device to
acknowledge the deflate operation): this allows the cleaner to be
used to release memory when an allocation fails and waiting is
permitted, and makes the cleaner more robust in out-of-memory
conditions because it can operate on stack-allocated memory
instead of requiring heap allocations.
Make the deflate function more robust against allocation failures,
and retry a deflate operation via a kernel timer if not all
requested pages could be deflated (e.g. due to allocation
failures).