Page cache: avoid modifying in-cache pages for the program ELF
When the user program is faulted-in on-demand, relevant pages are
put in the page cache; if the program code modifies this memory
(e.g. when writing to static variables), those changes are made to
the page cache, and even though they are not persisted to disk
(i.e. the executable file in the filesystem is not modified), they
appear when the executable file is read via standard syscalls, as
exemplified by the following code:
```
static int static_var;
void *b1 = malloc(100000);
int fd = open("executable_file", O_RDONLY);
int size = read(fd, b1, 100000);
static_var++;
void *b2 = malloc(100000);
lseek(fd, SEEK_SET, 0);
size = read(fd, b2, size);
printf("size %d memcmp %d\n", size, memcmp(b1, b2, size));
```
When the above code is run, the memcmp() function returns a
non-zero value, which indicates that the assignment to the static
variable is reflected in the in-cache memory backed by the
executable file.
Fix the above issue by adding a `private` boolean parameter to the
pagecache_get_page() and pagecache_get_page_if_filled() functions:
when the argument value is true, "steal" the page from the cache so
that it becomes "owned" by the user program; this allows any writes
to page memory by the program to not be reflected in any
file-backed pages, while avoiding the overhead of the copy-on-write
mechanism.