pagecache_get_page(): request memory cleaning on allocation failure
This makes on-demand retrieval of file-backed pages more robust in
low-memory conditions, e.g. it allows purging existing pages from
the cache in order to be able to create a new file-backed page to be accessed
by the user program (similarly to what is done when creating a new
page for anonymous mappings).