TLB shootdown: enforce CPU synchronization when necessary
The unmap_and_free_phys() function uses a range_handler to
deallocate physical memory pages as they are encountered during
page table traversal, and then calls page_invalidate_sync() to
perform a TLB shootdown. Between physical memory deallocation and
PTE invalidation, another CPU can potentially re-allocate the page
just freed and use it for other purposes while it is still
referenced in the TLB as associated to the mapping being torn down;
this can cause memory corruption issues. The same issue affects
the pagecache_node_unmap_pages() function.
A separate but similar issue affects the pagecache code that scans
shared file-backed mappings: when it finds a dirty page, it clears
the dirty flag, and then invokes a TLB shootdown; between the
clearing of the dirty flag and the PTE invalidation, another CPU
can potentially write to the mapped page and fail to re-set the
dirty flag in the PTE (because this flag may already be set in
the local TLB), causing the write to fail to be persisted on disk.

Fix the first issue by ensuring that a physical memory page is
deallocated only after TLB shootdown is performed. Try to do the
shootdown asynchronously (i.e. avoid waiting for the other CPUs to
do the invalidation) if possible, and resort to a synchronous
shootdown on memory allocation failure (the
unmap_and_free_phys_sync() and pagecache_node_unmap_pages_sync()
functions operate on stack memory only). Amend the TLB shoodown
code to implement a CPU "rendezvous" when a synchronous shootdown
is requested: the requesting CPU sends an IPI, waits for the other
CPUs to join the rendezvous, executes the completion closure
associated to the flush entry, then releases the other CPUs, which
can then leave the rendezvous.

Fix the second issue by invoking a synchronous TLB shootdown and
clearing the dirty flag during the CPU rendezvous.

Opportunistically amend the pagecache node data structure to
handle both shared and private mappings (in preparation for a
future commit which will implement scanning of private mappings),
using an embedded struct rangemap instead of a pointer to a
separately allocated struct, and adding proper locking.