x86/vsyscall: access vsyscall page via kernel virtual address
When initializing the vsyscall page, use its kernel virtual address
instead of its user virtual address: this allows avoiding the
subsequent call to update_map_flags() to make the page read-only,
and will allow the kernel to boot properly when running on the PVM
hypervisor (which does not permit the guest to dereference the
vsyscall user address).
Note: the vsyscall interface is still not supported on PVM, the
user program will get a segmentation fault if it tries to access
the vsyscall page.