Tagged memory: remove VA_TAG_BASE and tag() definitions
The tag() function is not needed in boot and kernel code, it's only
needed in the userspace implementation of the tagged allocator;
therefore, remove this function (and the VA_TAG_BASE symbol which
was only used in this function). This allows moving the KMEM_BASE
definition to kernel_machine.h, where it belongs (since it's a
kernel-specific definition).