Kernel virtual memory: add support for arbitrary address range
Instead of hard-coding the range of allowed kernel virtual
addresses, add a kvmem structure holding this range that is
initialized during boot. Since all kernel virtual memory has to be
allocated from this range, the range of linearly mapped memory
addresses is calculated during boot instead of being hardcoded.
This will allow adding support for the PVM hypervisor, which
imposes a restricted kernel address range on the guest.
Keep the default address values (for both the global range and the
linear mapping) to the previously defined hard-coded numbers.