x86: add cpuid and frame_return paravirtualized operations
When running with the PVM hypervisor, cpuid() and frame_return()
need to be implemented differently compared to when running on a
bare-metal (or equivalent) machine. Add an x86_pv_ops structure
that holds pointers to these functions, so that the appropriate
implementation can be selected during boot.
Actual PVM support will be added in a future commit.