virtio/net: fix processing of first buffer of merged RX buffers
When the virtio net feature that allows merging of RX buffers
is enabled, buffers belonging to a merged set must be processed in
strict sequential order; to achieve this, the seqno field of the
vnet_rx structure is used to prevent different CPUS from processing
RX buffers simultaneously. However, part of the code that processes
the first buffer of a merged set is being executed outside the
critical section, which can cause corruption of the relevant pbuf
struct. This issue can manifest itself as failure to process
incoming network traffic.
Fix the above issue by introducing the vnet_rx_hdr_remove() helper
function, which is called to process both the first buffer of a
merged set (inside the critical section), and the only buffer of a
received packet when merging is not being done.

Closes #2099.