virtio/net: add memory barriers when processing merged RX buffers
When the virtio net feature that allows merging of RX buffers is
enabled, the seqno field of the vnet_rx structure is used to
prevent different CPUS from processing simultaneously RX buffers
belonging to the same merged set.
Add memory barriers at the beginning and end of the critical
section, in order to prevent memory read/write reordering from
causing different CPUs to run parts of the critical section at the
same time.