Memory fences: add smp_read_barrier() and smp_write_barrier()
These macros are used when memory read/write ordering needs to be
enforced in order for different CPUs to have a consistent view of
shared memory. They can expand to an actual memory fence (on
architectures with a weakly ordered memory model), or to a simple
compiler barrier (on architectures with strong memory ordering).
The use of these macros instead of read_barrier() and
write_barrier() (which always insert a memory fence) creates more
efficient code for architectures (such as x86) with strong memory
ordering.
Replace calls to read/write_barrier() with calls to
smp_read/write_barrier() when memory is not used for device I/O;
remove barriers when not needed; in spin_unlock(), replace compiler
barrier with SMP write barrier, otherwise code is not safe with
weakly ordered memory model.