CircleCI image: switch from cimg/go:1.20.0 to cimg/go:1.25.0
This fixes the following error in the CircleCI job that runs
`make test-noaccel`:
```
github.com/nanovms/ops/lepton imports
        github.com/schollz/progressbar/v3 imports
        golang.org/x/crypto/ssh/terminal imports
        golang.org/x/term imports
        golang.org/x/sys/unix imports
        slices: package slices is not in GOROOT (/usr/local/go/src/slices)
note: imported by a module that requires go 1.23
```
In the build-riscv job, the CircleCI image is not upgraded because
there is an issue with the futtexrobust runtime test that fails
when the program is built with gcc 13.

In addition, update the go.mod file for Go tests.