Add klib to read userdata from instance metadata service
A new klib 'userdata' is added. This will attempt to read the 'userdata'
from various cloud providers and either use it to set environment
variables, or write it to a file.

This can be used to separate preparing images and setting
instance-specific data.

Cloud providers typically have IMDS (Instance Metadata service) that is
ONLY reachable from within the instance on 169.254.169.254.