filesystem: remove ".." attribute from non-directory entries
When support for hard links is added, a directory entry that is not a
directory itself may have more than one parent, thus the ".." attribute is
appropriate only for directories.
In filesystem code, only assign the ".." attribute to directories; in the
inotify code, assign the ".." attribute to non-directory entries when
needed to be able to send events to watched directories.
To preperly handle IN_MODIFY events when truncating a file, move inotify
calls from the filesystem layer to the Unix layer.
Amend the unveil syscall code to be able to walk up a file path when the
".." attribute is not present for non-directory entries.