pagecache node fetch: fix handling of queries past the end of a file
If pagecache_node_fetch_pages() is called with a range argument that
overlaps with the last cache page for the file node but extends beyond the
file length, the scatter-gather list is filled with data past the end of
the file. This can cause sendfile() to return non-zero values when the
input file offset is at the end of the file.
Instead of clamping the read range to the file length from
pagecache_read_sg(), revise the pp_handler implementation for both
pagecache_read_sg() and pagecache_node_fetch_pages() so that only the data
actually available in the file being read is processed; in
pagecache_fetch_pp_handler(), remove a bogus refcount increment; in
addition, amend pagecache_node_fetch_internal() so that it only reads the
node length when the node is locked.
In the sendfile runtime test program, add a test case that would have
failed without this fix.